move diagnostic tasks to common path
